src/main/java/com/project/_2023_01_26/mockito/code/reactor/Thermometer.java
package com.project._2023_01_26.mockito.code.reactor;

public class Thermometer {

    private double temperature;
    private Sensor sensor;

    public double getTemperature() {
        if (sensor.isBlocked()) {
            throw new RuntimeException("Sensor is blocked");
        }
        return temperature;
    }

    public void setTemperature(double temperature) {
        this.temperature = temperature;
    }

    public Sensor getSensor() {
        return sensor;
    }

    public void setSensor(Sensor sensor) {
        this.sensor = sensor;
    }
}